+++ src/App.js
@@ -3,6 +3,10 @@ import { BrowserRouter as Router,Route } from "react-router-dom";
 
 import CreateRouteSharing from './components/CreateRouteSharing'
 import RoutesMap from './components/CreateRouteSharing/RoutesMap'
+import SidenavPushMenu from './components/SidenavPushMenu'
+import SidenavPushNearbyUsers from './components/SidenavPushNearbyUsers'
+import SidenavPushSameWayNearbyUsers from './components/SidenavPushSameWayNearbyUsers'
+import SidenavPushSearchLocationNearbyUsers from './components/SidenavPushSearchLocationNearbyUsers'
 // Firebase.
 import * as firebase from 'firebase';
 // import { RoutePages } from './RoutePages'
@@ -137,18 +141,19 @@ class App extends React.Component {
             <React.Fragment>
                 {this.state.isSignedIn ? (
                     <Router>
-                        <Route path="/" exact render={ () => <Maps {...state} >
-                            {/* <h1> Hello.  {firebaseApp.auth().currentUser.displayName} You are now signed In! </h1> */}
-
-                            <a 
-                            className="mm-listitem__text-menu" 
-                            onClick={() => firebaseApp.auth().signOut()} 
-                            >ออกจากระบบ</a>
-
-                        </Maps>
+                        <Route path="/" exact render={ () => <Maps {...state} />
                         } />
                         <Route path="/create_route_sharing" render={() => <CreateRouteSharing {...state} />} />
                         <Route path="/routes_map" render={() => <RoutesMap {...state} /> } />
+                        <Route path="/menu" render={() => <SidenavPushMenu {...state} >
+                        <a 
+                            className="mm-listitem__text-menu" 
+                            onClick={() => firebaseApp.auth().signOut()} 
+                            >ออกจากระบบ</a>
+                        </SidenavPushMenu> } />
+                        <Route path="/near_by_users" render={() => <SidenavPushNearbyUsers {...state} /> } />
+                        <Route path="/same_way_near_by_users" render={() => <SidenavPushSameWayNearbyUsers {...state} /> } />
+                        <Route path="/search_location_near_by_users" render={() => <SidenavPushSearchLocationNearbyUsers {...state} /> } />
                         {/* <RoutePages /> */}
                     </Router>
                 )
