+++ src/components/FormPageMore/styles/input-style.css
@@ -1,4 +1,4 @@
-.form-control-plaintext-sidenav-push-near-by-users {
+.form-control-plaintext-form-page-more {
     display: block;
     /* width: 22%; */
     /* padding-top: .375rem; */
