+++ src/components/FormPageMore/styles/mySidenav.css
@@ -1,6 +1,6 @@
-.overlay-same-way-nearby-users {
+.overlay-form-page-more {
   height: 100%;
-  width: 0%;
+  width: 100%;
   position: fixed;
   z-index: 1;
   top: 0;
@@ -18,7 +18,7 @@
   margin-top: 30px;
 }
 
-.overlay-same-way-nearby-users a {
+.overlay-form-page-more a {
   padding: 8px;
   text-decoration: none;
   font-size: 36px;
@@ -27,11 +27,11 @@
   transition: 0.3s;
 }
 
-.overlay-same-way-nearby-users a:hover, .overlay-same-way-nearby-users a:focus {
+.overlay-form-page-more a:hover, .overlay-form-page-more a:focus {
   color: #f1f1f1;
 }
 
-.overlay-same-way-nearby-users .closebtn-same-way-nearby-users {
+.overlay-form-page-more .closebtn-form-page-more {
   position: absolute;
   top: 2px;
   left: 5px;
@@ -40,9 +40,9 @@
 }
 
 @media screen and (max-height: 450px) {
-  .overlay-same-way-nearby-users {overflow-y: auto;}
-  .overlay-same-way-nearby-users a {font-size: 20px}
-  .overlay-same-way-nearby-users .closebtn-same-way-nearby-users {
+  .overlay-form-page-more {overflow-y: auto;}
+  .overlay-form-page-more a {font-size: 20px}
+  .overlay-form-page-more .closebtn-form-page-more {
   font-size: 40px;
   top: 15px;
   right: 35px;
