--- src/components/FromChat/ListChat.js
@@ -1,17 +1,0 @@
-import React from 'react'
-
-const ListChat = (props) => {
-
-    return (
-        <div style={{ marginTop:10 }}>
-            {
-                props.listMsg.map( msg =>{
-                    return <p key = { msg.key }>{ msg.message }</p>
-                }) 
-            }
-        </div>        
-    )
-
-}
-
-export default ListChat
