--- src/components/Maps/SameWayNearbyUsers/styles/MuiSvgIcon.css
@@ -1,35 +1,0 @@
-
-.MuiSvgIcon-root-18 {
-    fill: currentColor;
-    width: 1em;
-    height: 1em;
-    display: inline-block;
-    font-size: 24px;
-    transition: fill 200ms cubic-bezier(0.4, 0, 0.2, 1) 0ms;
-    user-select: none;
-    flex-shrink: 0;
-  }
-  .MuiSvgIcon-colorPrimary-19 {
-    color: rgb(22, 113, 216);
-  }
-  .MuiSvgIcon-colorSecondary-20 {
-    color: #f50057;
-  }
-  .MuiSvgIcon-colorAction-21 {
-    color: rgba(0, 0, 0, 0.54);
-  }
-  .MuiSvgIcon-colorError-22 {
-    color: #f44336;
-  }
-  .MuiSvgIcon-colorDisabled-23 {
-    color: rgba(0, 0, 0, 0.26);
-  }
-  .MuiSvgIcon-fontSizeInherit-24 {
-    font-size: inherit;
-  }
-  .MuiSvgIcon-fontSizeSmall-25 {
-    font-size: 20px;
-  }
-  .MuiSvgIcon-fontSizeLarge-26 {
-    font-size: 35px;
-  }
