--- src/components/Maps/SameWayNearbyUsers/styles/MuiTouchRipple.css
@@ -1,87 +1,0 @@
-.MuiTouchRipple-root-27 {
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    display: block;
-    z-index: 0;
-    position: absolute;
-    overflow: hidden;
-    border-radius: inherit;
-    pointer-events: none;
-}
-
-.MuiTouchRipple-ripple-28 {
-    top: 0;
-    left: 0;
-    width: 50px;
-    height: 50px;
-    opacity: 0;
-    position: absolute;
-}
-
-.MuiTouchRipple-rippleVisible-29 {
-    opacity: 0.3;
-    transform: scale(1);
-    animation: mui-ripple-enter 550ms cubic-bezier(0.4, 0, 0.2, 1);
-    animation-name: '$mui-ripple-enter';
-}
-
-.MuiTouchRipple-ripplePulsate-30 {
-    animation-duration: 200ms;
-}
-
-.MuiTouchRipple-child-31 {
-    width: 100%;
-    height: 100%;
-    opacity: 1;
-    display: block;
-    border-radius: 50%;
-    background-color: currentColor;
-}
-
-.MuiTouchRipple-childLeaving-32 {
-    opacity: 0;
-    animation: mui-ripple-exit 550ms cubic-bezier(0.4, 0, 0.2, 1);
-    animation-name: '$mui-ripple-exit';
-}
-
-.MuiTouchRipple-childPulsate-33 {
-    top: 0;
-    left: 0;
-    position: absolute;
-    animation: mui-ripple-pulsate 2500ms cubic-bezier(0.4, 0, 0.2, 1) 200ms infinite;
-    animation-name:' $mui-ripple-pulsate';
-}
-
-@-webkit-keyframes mui-ripple-enter {
-    0% {
-        opacity: 0.1;
-        transform: scale(0);
-    }
-    100% {
-        opacity: 0.3;
-        transform: scale(1);
-    }
-}
-
-@-webkit-keyframes mui-ripple-exit {
-    0% {
-        opacity: 1;
-    }
-    100% {
-        opacity: 0;
-    }
-}
-
-@-webkit-keyframes mui-ripple-pulsate {
-    0% {
-        transform: scale(1);
-    }
-    50% {
-        transform: scale(0.92);
-    }
-    100% {
-        transform: scale(1);
-    }
-}
