--- src/components/Maps/SearchLocationNearbyUsers/styles/MuiButtonBase.css
@@ -1,28 +1,0 @@
-
-.MuiButtonBase-root-15 {
-    color: inherit;
-    border: 0;
-    margin: 0;
-    cursor: pointer;
-    display: inline-flex;
-    outline: none;
-    padding: 0;
-    position: relative;
-    align-items: center;
-    user-select: none;
-    border-radius: 0;
-    vertical-align: middle;
-    justify-content: center;
-    -moz-appearance: none;
-    text-decoration: none;
-    background-color: transparent;
-    -webkit-appearance: none;
-    -webkit-tap-highlight-color: transparent;
-  }
-  .MuiButtonBase-root-15::-moz-focus-inner {
-    border-style: none;
-  }
-  .MuiButtonBase-root-15.MuiButtonBase-disabled-16 {
-    cursor: default;
-    pointer-events: none;
-  }
