--- src/components/Maps/SearchLocationNearbyUsers/styles/MuiFab.css
@@ -1,106 +1,0 @@
-
-.MuiFab-root-5 {
-    color: rgba(0, 0, 0, 0.87);
-    width: 56px;
-    height: 56px;
-    padding: 0;
-    font-size: 0.875rem;
-    min-width: 0;
-    box-sizing: border-box;
-    min-height: 36px;
-    transition: background-color 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms,box-shadow 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms,border 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms;
-    box-shadow: 0px 3px 5px -1px rgba(0,0,0,0.2),0px 6px 10px 0px rgba(0,0,0,0.14),0px 1px 18px 0px rgba(0,0,0,0.12);
-    font-family: "Roboto", "Helvetica", "Arial", sans-serif;
-    font-weight: 500;
-    line-height: 1.75;
-    border-radius: 50%;
-    letter-spacing: 0.02857em;
-    text-transform: uppercase;
-    background-color: #e0e0e0;
-  }
-  .MuiFab-root-5:active {
-    box-shadow: 0px 7px 8px -4px rgba(0,0,0,0.2),0px 12px 17px 2px rgba(0,0,0,0.14),0px 5px 22px 4px rgba(0,0,0,0.12);
-  }
-  .MuiFab-root-5.MuiFab-focusVisible-10 {
-    box-shadow: 0px 3px 5px -1px rgba(0,0,0,0.2),0px 6px 10px 0px rgba(0,0,0,0.14),0px 1px 18px 0px rgba(0,0,0,0.12);
-  }
-  .MuiFab-root-5:hover {
-    text-decoration: none;
-    background-color: #d5d5d5;
-  }
-  .MuiFab-root-5.MuiFab-disabled-11 {
-    color: rgba(0, 0, 0, 0.26);
-    box-shadow: none;
-    background-color: rgba(0, 0, 0, 0.12);
-  }
-  @media (hover: none) {
-    .MuiFab-root-5:hover {
-      background-color: #e0e0e0;
-    }
-  }
-  .MuiFab-root-5:hover.MuiFab-disabled-11 {
-    background-color: rgba(0, 0, 0, 0.12);
-  }
-  .MuiFab-label-6 {
-    width: 100%;
-    display: inherit;
-    align-items: inherit;
-    justify-content: inherit;
-  }
-  .MuiFab-primary-7 {
-    color: #fff;
-    background-color: #1D385A;
-  }
-  .MuiFab-primary-7:hover {
-    background-color: rgb(20, 45, 75);
-  }
-  @media (hover: none) {
-    .MuiFab-primary-7:hover {
-      background-color: #1D385A;
-    }
-  }
-  .MuiFab-secondary-8 {
-    color: #fff;
-    background-color: #f50057;
-  }
-  .MuiFab-secondary-8:hover {
-    background-color: #c51162;
-  }
-  @media (hover: none) {
-    .MuiFab-secondary-8:hover {
-      background-color: #f50057;
-    }
-  }
-  .MuiFab-extended-9 {
-    width: auto;
-    height: 48px;
-    padding: 0 16px;
-    min-width: 48px;
-    min-height: auto;
-    border-radius: 24px;
-  }
-  .MuiFab-extended-9.MuiFab-sizeSmall-13 {
-    width: auto;
-    height: 34px;
-    padding: 0 8px;
-    min-width: 34px;
-    border-radius: 17px;
-  }
-  .MuiFab-extended-9.MuiFab-sizeMedium-14 {
-    width: auto;
-    height: 40px;
-    padding: 0 16px;
-    min-width: 40px;
-    border-radius: 20px;
-  }
-  .MuiFab-colorInherit-12 {
-    color: inherit;
-  }
-  .MuiFab-sizeSmall-13 {
-    width: 40px;
-    height: 40px;
-  }
-  .MuiFab-sizeMedium-14 {
-    width: 36px;
-    height: 36px;
-  }
