--- src/components/Maps/SidenavPushMenu/styles/mySidenav.css
@@ -1,51 +1,0 @@
-.overlay-menu {
-  height: 0%;
-  width: 100%;
-  position: fixed;
-  z-index: 1;
-  top: 0;
-  left: 0;
-  background-color: #274D7D;
-  overflow-y: hidden;
-  transition: 0.5s;
-}
-
-.overlay-content {
-  position: relative;
-  top: 25%;
-  width: 100%;
-  text-align: center;
-  margin-top: 30px;
-}
-
-.overlay-menu a {
-  padding: 8px;
-  text-decoration: none;
-  font-size: 36px;
-  color: #ffffff;
-  display: block;
-  transition: 0.3s;
-}
-
-.overlay-menu a:hover, .overlay-menu a:focus {
-  color: #f1f1f1;
-}
-
-.overlay-menu .closebtn-menu {
-  position: absolute;
-  top: 2px;
-  left: 5px;
-  font-size: 25px;
-  z-index: 10;
-}
-
-@media screen and (max-height: 450px) {
-  .overlay-menu {overflow-y: auto;}
-  .overlay-menu a {font-size: 20px}
-  .overlay-menu .closebtn-menu {
-  font-size: 40px;
-  top: 15px;
-  right: 35px;
-  }
-}
-
