--- src/components/Maps/SidenavPushSearchLocationNearbyUsers/styles/mySidenav.css
@@ -1,51 +1,0 @@
-.overlay-search-location-nearby-users {
-  height: 100%;
-  width: 0%;
-  position: fixed;
-  z-index: 1;
-  top: 0;
-  left: 0;
-  background-color: #274D7D;
-  overflow-x: hidden;
-  transition: 0.5s;
-}
-
-.overlay-content {
-  position: relative;
-  top: 25%;
-  width: 100%;
-  text-align: center;
-  margin-top: 30px;
-}
-
-.overlay-search-location-nearby-users a {
-  padding: 8px;
-  text-decoration: none;
-  font-size: 36px;
-  color: #ffffff;
-  display: block;
-  transition: 0.3s;
-}
-
-.overlay-search-location-nearby-users a:hover, .overlay-search-location-nearby-users a:focus {
-  color: #f1f1f1;
-}
-
-.overlay-search-location-nearby-users .closebtn-search-location-nearby-users {
-  position: absolute;
-  top: 2px;
-  left: 5px;
-  font-size: 25px;
-  z-index: 10;
-}
-
-@media screen and (max-height: 450px) {
-  .overlay-search-location-nearby-users {overflow-y: auto;}
-  .overlay-search-location-nearby-users a {font-size: 20px}
-  .overlay-search-location-nearby-users .closebtn-search-location-nearby-users {
-  font-size: 40px;
-  top: 15px;
-  right: 35px;
-  }
-}
-
