+++ src/components/SidenavPushMenu/styles/mySidenav.css
@@ -1,5 +1,5 @@
-.overlay {
-  height: 0%;
+.overlay-menu {
+  height: -webkit-fill-available;
   width: 100%;
   position: fixed;
   z-index: 1;
@@ -18,7 +18,7 @@
   margin-top: 30px;
 }
 
-.overlay a {
+.overlay-menu a {
   padding: 8px;
   text-decoration: none;
   font-size: 36px;
@@ -27,11 +27,11 @@
   transition: 0.3s;
 }
 
-.overlay a:hover, .overlay a:focus {
+.overlay-menu a:hover, .overlay-menu a:focus {
   color: #f1f1f1;
 }
 
-.overlay .closebtn {
+.overlay-menu .closebtn-menu {
   position: absolute;
   top: 2px;
   left: 5px;
@@ -40,9 +40,9 @@
 }
 
 @media screen and (max-height: 450px) {
-  .overlay {overflow-y: auto;}
-  .overlay a {font-size: 20px}
-  .overlay .closebtn {
+  .overlay-menu {overflow-y: auto;}
+  .overlay-menu a {font-size: 20px}
+  .overlay-menu .closebtn-menu {
   font-size: 40px;
   top: 15px;
   right: 35px;
