--- src/components/items/Font.js
@@ -1,22 +1,0 @@
-// // JasmineUPC
-
-// import React from 'react'
-// import { withStyles } from '@material-ui/core/styles'
-// const bts = theme => ({
-//     FontIII: {
-//         fontStyle:'JasmineUPC'
-//      },
-// })
-
-// class HeadTop extends React.Component {
-
-//     render() {
-//         return (
-//             <header className={this.props.classes.FontIII} >
-
-//             </header>
-
-//         )
-//     }
-// }
-// export default withStyles(bts)(HeadTop)
