--- src/components/items/LoginForm/HeadText.js
@@ -1,36 +1,0 @@
-import React from 'react'
-import { withStyles } from '@material-ui/core/styles'
-// import color from '@material-ui/core/colors/blueGrey';
-
-const bts = theme => ({
-    centerrI: {
-        marginTop: '9%',
-        width: '95%',
-        height: '1px',
-        color: 'white',
-        backgroundColor: 'white'
-    },
-    centerrII: {
-        margin: '22%',
-        width: '60%',
-        color: 'white',
-        marginTop:'9%'
-        
-    },
-    }
-)
-
-class CheckNum extends React.Component {
-
-    render() {
-        return (
-            <React.Fragment>
-
-                <hr className={this.props.classes.centerrI}></hr>
-                <h3 className={this.props.classes.centerrII}>{this.props.text}</h3>
-
-            </React.Fragment>
-        )
-    }
-}
-export default withStyles(bts)(CheckNum)
